feat: RoomRepository 내 개별 숙소 조회 기능 구현
